html/browsers/history/the-history-interface/history_replacestate_url_emptystring_with_base.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>history.replaceState() with an empty string URL and base URL different from document's URL</title>
<link rel="help" href="https://github.com/whatwg/html/issues/9343">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<base href="/the-base">

<script>
"use strict";

test(() => {
  history.replaceState(null, null, "");
  assert_equals(location.pathname, "/the-base", "location.pathname");
  assert_true(location.href.endsWith("/the-base"), "location.href");
  assert_true(document.URL.endsWith("/the-base"), "document.URL");
});
</script>
